TimeOutTimeIntervalExample
