+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './src/components/ImageComponentExample';
+import App from './src/components/TimeoutTimeIntervalExample';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
